server/requirements.txt
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
rich==13.6.0
urllib3==2.1.0
Werkzeug==3.0.1
